If cast diff from 'long' to 'int' forcibly, it will overflow when 'periodicityType' is 'TOP_OF_HOUR' and 'cleanHistoryOnStart' property is set to true.

+++ logback-core/src/main/java/ch/qos/logback/core/rolling/helper/RollingCalendar.java
@@ -196,7 +196,7 @@ public class RollingCalendar extends GregorianCalendar {
             return diff / MILLIS_IN_ONE_MINUTE;
         case TOP_OF_HOUR:
 
-            return (int) diff / MILLIS_IN_ONE_HOUR;
+            return diff / MILLIS_IN_ONE_HOUR;
         case TOP_OF_DAY:
             return diff / MILLIS_IN_ONE_DAY;
         case TOP_OF_WEEK:
